Trivial changes

+++ collections_common.h
@@ -6,7 +6,7 @@ namespace nativa
 {
 	namespace collections
 	{
-		using index_type = long;
+		using index_type = size_t;
     }
 }
